cleanup unused code, break out transforms
