+++ deeplab/yamls/mcloud_run.yaml
@@ -66,7 +66,7 @@ parameters:
       # project: # Name of WandB project
       # group:   # Name of WandB group
 
-  # null for baseline or for recipe, either ["mild", "medium", "spicy"] in order of increasing training time and accuracy
+  # null for baseline or for recipe, either ["mild", "medium", "hot"] in order of increasing training time and accuracy
   recipe_name: null
 
   # Updated parameters for mild recipe
@@ -88,3 +88,4 @@ parameters:
 
   # Load checkpoint parameters
   load_path: null # e.g. './ckpt/latest-rank{rank}.pt' (local) or 's3://mybucket/mydir/ckpt/latest-rank{rank}.pt' (remote)
+
